+++ geo-types/src/coordinate.rs
@@ -1,12 +1,14 @@
-use crate::{coord, CoordNum, Point};
-
+use crate::{coord, CoordNum, Measure, NoValue, Point, ZCoord};
 #[cfg(any(feature = "approx", test))]
 use approx::{AbsDiffEq, RelativeEq, UlpsEq};
+use num_traits::Zero;
+use std::fmt::Debug;
+use std::ops::{Add, Div, Mul, Neg, Sub};
 
-/// A lightweight struct used to store coordinates on the 2-dimensional
-/// Cartesian plane.
+/// A generic struct used to store a single coordinate with optional
+/// 3D (Z) and measurement support.
 ///
-/// Unlike `Point` (which in the future may contain additional information such
+/// Unlike [`Point`] (which in the future may contain additional information such
 /// as an envelope, a precision model, and spatial reference system
 /// information), a `Coordinate` only contains ordinate values and accessor
 /// methods.
@@ -25,12 +27,43 @@ use approx::{AbsDiffEq, RelativeEq, UlpsEq};
 /// [vector space]: //en.wikipedia.org/wiki/Vector_space
 #[derive(Eq, PartialEq, Clone, Copy, Debug, Hash, Default)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct Coordinate<T: CoordNum> {
+pub struct Coordinate<T: CoordNum, Z: ZCoord = NoValue, M: Measure = NoValue> {
     pub x: T,
     pub y: T,
+    pub z: Z,
+    pub m: M,
 }
 
+impl<T: CoordNum, Z: ZCoord, M: Measure> Coordinate<T, Z, M> {
+    /// Create a new instance of a coordinate.
+    /// **ATTENTION** Use [`coord!`] macro or one of the [`from`] methods instead.
+    #[inline]
+    #[doc(hidden)]
+    pub fn new__(x: T, y: T, z: Z, m: M) -> Self {
+        Self { x, y, z, m }
+    }
+}
+
+/// A lightweight struct used to store coordinates on the 2-dimensional
+/// Cartesian plane together with a Measure value of the same type.
+///
+/// See also [Coordinate]
+pub type CoordinateM<T, M = T> = Coordinate<T, NoValue, M>;
+
+/// A lightweight struct used to store coordinates on the 3-dimensional
+/// Cartesian plane.
+///
+/// See also [Coordinate]
+pub type Coordinate3D<T> = Coordinate<T, T, NoValue>;
+
+/// A lightweight struct used to store coordinates on the 3-dimensional
+/// Cartesian plane together with a Measure value of the same type.
+///
+/// See also [Coordinate]
+pub type Coordinate3DM<T, M = T> = Coordinate<T, T, M>;
+
 impl<T: CoordNum> From<(T, T)> for Coordinate<T> {
+    #[inline]
     fn from(coords: (T, T)) -> Self {
         coord! {
             x: coords.0,
@@ -40,6 +73,7 @@ impl<T: CoordNum> From<(T, T)> for Coordinate<T> {
 }
 
 impl<T: CoordNum> From<[T; 2]> for Coordinate<T> {
+    #[inline]
     fn from(coords: [T; 2]) -> Self {
         coord! {
             x: coords[0],
@@ -48,28 +82,28 @@ impl<T: CoordNum> From<[T; 2]> for Coordinate<T> {
     }
 }
 
-impl<T: CoordNum> From<Point<T>> for Coordinate<T> {
-    fn from(point: Point<T>) -> Self {
-        coord! {
-            x: point.x(),
-            y: point.y(),
-        }
+impl<T: CoordNum, Z: ZCoord, M: Measure> From<Point<T, Z, M>> for Coordinate<T, Z, M> {
+    #[inline]
+    fn from(point: Point<T, Z, M>) -> Self {
+        point.0
     }
 }
 
-impl<T: CoordNum> From<Coordinate<T>> for (T, T) {
-    fn from(coord: Coordinate<T>) -> Self {
+impl<T: CoordNum, Z: ZCoord, M: Measure> From<Coordinate<T, Z, M>> for (T, T) {
+    #[inline]
+    fn from(coord: Coordinate<T, Z, M>) -> Self {
         (coord.x, coord.y)
     }
 }
 
-impl<T: CoordNum> From<Coordinate<T>> for [T; 2] {
-    fn from(coord: Coordinate<T>) -> Self {
+impl<T: CoordNum, Z: ZCoord, M: Measure> From<Coordinate<T, Z, M>> for [T; 2] {
+    #[inline]
+    fn from(coord: Coordinate<T, Z, M>) -> Self {
         [coord.x, coord.y]
     }
 }
 
-impl<T: CoordNum> Coordinate<T> {
+impl<T: CoordNum, Z: ZCoord, M: Measure> Coordinate<T, Z, M> {
     /// Returns a tuple that contains the x/horizontal & y/vertical component of the coordinate.
     ///
     /// # Examples
@@ -86,12 +120,11 @@ impl<T: CoordNum> Coordinate<T> {
     /// assert_eq!(y, 116.34);
     /// assert_eq!(x, 40.02f64);
     /// ```
+    #[inline]
     pub fn x_y(&self) -> (T, T) {
         (self.x, self.y)
     }
 }
-
-use std::ops::{Add, Div, Mul, Neg, Sub};
 
 /// Negate a coordinate.
 ///
@@ -106,16 +139,21 @@ use std::ops::{Add, Div, Mul, Neg, Sub};
 /// assert_eq!(q.x, -p.x);
 /// assert_eq!(q.y, -p.y);
 /// ```
-impl<T> Neg for Coordinate<T>
+impl<T, Z, M> Neg for Coordinate<T, Z, M>
 where
     T: CoordNum + Neg<Output = T>,
+    Z: ZCoord + Neg<Output = Z>,
+    M: Measure + Neg<Output = M>,
 {
     type Output = Self;
 
+    #[inline]
     fn neg(self) -> Self {
         coord! {
             x: -self.x,
             y: -self.y,
+            z: -self.z,
+            m: -self.m,
         }
     }
 }
@@ -134,13 +172,16 @@ where
 /// assert_eq!(sum.x, 2.75);
 /// assert_eq!(sum.y, 5.0);
 /// ```
-impl<T: CoordNum> Add for Coordinate<T> {
+impl<T: CoordNum, Z: ZCoord, M: Measure> Add for Coordinate<T, Z, M> {
     type Output = Self;
 
+    #[inline]
     fn add(self, rhs: Self) -> Self {
         coord! {
             x: self.x + rhs.x,
             y: self.y + rhs.y,
+            z: self.z + rhs.z,
+            m: self.m + rhs.m,
         }
     }
 }
@@ -159,13 +200,16 @@ impl<T: CoordNum> Add for Coordinate<T> {
 /// assert_eq!(diff.x, 0.25);
 /// assert_eq!(diff.y, 0.);
 /// ```
-impl<T: CoordNum> Sub for Coordinate<T> {
+impl<T: CoordNum, Z: ZCoord, M: Measure> Sub for Coordinate<T, Z, M> {
     type Output = Self;
 
+    #[inline]
     fn sub(self, rhs: Self) -> Self {
         coord! {
             x: self.x - rhs.x,
             y: self.y - rhs.y,
+            z: self.z - rhs.z,
+            m: self.m - rhs.m,
         }
     }
 }
@@ -183,13 +227,21 @@ impl<T: CoordNum> Sub for Coordinate<T> {
 /// assert_eq!(q.x, 5.0);
 /// assert_eq!(q.y, 10.0);
 /// ```
-impl<T: CoordNum> Mul<T> for Coordinate<T> {
+impl<T, Z, M> Mul<T> for Coordinate<T, Z, M>
+where
+    T: CoordNum,
+    Z: ZCoord + Mul<T, Output = Z>,
+    M: Measure + Mul<T, Output = M>,
+{
     type Output = Self;
 
+    #[inline]
     fn mul(self, rhs: T) -> Self {
         coord! {
             x: self.x * rhs,
             y: self.y * rhs,
+            z: self.z * rhs,
+            m: self.m * rhs,
         }
     }
 }
@@ -207,18 +259,25 @@ impl<T: CoordNum> Mul<T> for Coordinate<T> {
 /// assert_eq!(q.x, 1.25);
 /// assert_eq!(q.y, 2.5);
 /// ```
-impl<T: CoordNum> Div<T> for Coordinate<T> {
+impl<T, Z, M> Div<T> for Coordinate<T, Z, M>
+where
+    T: CoordNum,
+    Z: ZCoord + Div<T, Output = Z>,
+    M: Measure + Div<T, Output = M>,
+{
     type Output = Self;
 
+    #[inline]
     fn div(self, rhs: T) -> Self {
         coord! {
             x: self.x / rhs,
             y: self.y / rhs,
+            z: self.z / rhs,
+            m: self.m / rhs,
         }
     }
 }
 
-use num_traits::Zero;
 /// Create a coordinate at the origin.
 ///
 /// # Examples
@@ -232,21 +291,26 @@ use num_traits::Zero;
 /// assert_eq!(p.x, 0.);
 /// assert_eq!(p.y, 0.);
 /// ```
-impl<T: CoordNum> Coordinate<T> {
+impl<T: CoordNum, Z: ZCoord, M: Measure> Coordinate<T, Z, M> {
+    #[inline]
     pub fn zero() -> Self {
         coord! {
             x: T::zero(),
             y: T::zero(),
+            z: Z::zero(),
+            m: M::zero(),
         }
     }
 }
 
-impl<T: CoordNum> Zero for Coordinate<T> {
+impl<T: CoordNum, Z: ZCoord, M: Measure> Zero for Coordinate<T, Z, M> {
+    #[inline]
     fn zero() -> Self {
         Self::zero()
     }
+    #[inline]
     fn is_zero(&self) -> bool {
-        self.x.is_zero() && self.y.is_zero()
+        self.x.is_zero() && self.y.is_zero() && self.z.is_zero() && self.m.is_zero()
     }
 }
 
@@ -290,10 +354,12 @@ impl<T: CoordNum + UlpsEq> UlpsEq for Coordinate<T>
 where
     T::Epsilon: Copy,
 {
+    #[inline]
     fn default_max_ulps() -> u32 {
         T::default_max_ulps()
     }
 
+    #[inline]
     fn ulps_eq(&self, other: &Self, epsilon: T::Epsilon, max_ulps: u32) -> bool {
         T::ulps_eq(&self.x, &other.x, epsilon, max_ulps)
             && T::ulps_eq(&self.y, &other.y, epsilon, max_ulps)
@@ -309,6 +375,7 @@ where
 
     const DIMENSIONS: usize = 2;
 
+    #[inline]
     fn generate(generator: impl Fn(usize) -> Self::Scalar) -> Self {
         coord! {
             x: generator(0),
@@ -316,6 +383,7 @@ where
         }
     }
 
+    #[inline]
     fn nth(&self, index: usize) -> Self::Scalar {
         match index {
             0 => self.x,
@@ -324,6 +392,7 @@ where
         }
     }
 
+    #[inline]
     fn nth_mut(&mut self, index: usize) -> &mut Self::Scalar {
         match index {
             0 => &mut self.x,
@@ -342,6 +411,7 @@ where
 
     const DIMENSIONS: usize = 2;
 
+    #[inline]
     fn generate(mut generator: impl FnMut(usize) -> Self::Scalar) -> Self {
         coord! {
             x: generator(0),
@@ -349,6 +419,7 @@ where
         }
     }
 
+    #[inline]
     fn nth(&self, index: usize) -> Self::Scalar {
         match index {
             0 => self.x,
@@ -357,11 +428,44 @@ where
         }
     }
 
+    #[inline]
     fn nth_mut(&mut self, index: usize) -> &mut Self::Scalar {
         match index {
             0 => &mut self.x,
             1 => &mut self.y,
             _ => unreachable!(),
         }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    #[test]
+    fn test_coordinates() {
+        let p = coord! { x: 1.0, y: 2.0 };
+        assert_relative_eq!(p.x, 1.0);
+        assert_relative_eq!(p.y, 2.0);
+        assert_eq!(p.z, NoValue);
+        assert_eq!(p.m, NoValue);
+
+        let p = coord! { x: 1.0, y: 2.0, z: 3.0 };
+        assert_relative_eq!(p.x, 1.0);
+        assert_relative_eq!(p.y, 2.0);
+        assert_relative_eq!(p.z, 3.0);
+        assert_eq!(p.m, NoValue);
+
+        let p = coord! { x: 1.0, y: 2.0, m: 4_u8 };
+        assert_relative_eq!(p.x, 1.0);
+        assert_relative_eq!(p.y, 2.0);
+        assert_eq!(p.z, NoValue);
+        assert_eq!(p.m, 4_u8);
+
+        let p = coord! { x: 1_i32, y: 2_i32, z: 3_i32, m: 4.0_f64 };
+        assert_eq!(p.x, 1);
+        assert_eq!(p.y, 2);
+        assert_eq!(p.z, 3);
+        assert_relative_eq!(p.m, 4.0);
     }
 }