+++ geo-types/src/macros.rs
@@ -55,8 +55,22 @@ macro_rules! point {
 /// [`Coordinate`]: ./struct.Point.html
 #[macro_export]
 macro_rules! coord {
-    (x: $x:expr, y: $y:expr $(,)?) => {
-        $crate::Coordinate { x: $x, y: $y }
+    (x: $x:expr, y: $y:expr $(,)* ) => {
+        $crate::Coordinate::new__(
+            $x,
+            $y,
+            $crate::NoValue::default(),
+            $crate::NoValue::default(),
+        )
+    };
+    (x: $x:expr, y: $y:expr, z: $z:expr $(,)* ) => {
+        $crate::Coordinate::new__($x, $y, $z, $crate::NoValue::default())
+    };
+    (x: $x:expr, y: $y:expr, m: $m:expr $(,)* ) => {
+        $crate::Coordinate::new__($x, $y, $crate::NoValue::default(), $m)
+    };
+    (x: $x:expr, y: $y:expr, z: $z:expr, m: $m:expr $(,)* ) => {
+        $crate::Coordinate::new__($x, $y, $z, $m)
     };
 }
 
@@ -123,7 +137,7 @@ macro_rules! coord {
 /// [`LineString`]: ./line_string/struct.LineString.html
 #[macro_export]
 macro_rules! line_string {
-    () => { $crate::LineString(vec![]) };
+    () => { $crate::LineString::new(vec![]) };
     (
         $(( $($tag:tt : $val:expr),* $(,)? )),*
         $(,)?
@@ -138,7 +152,7 @@ macro_rules! line_string {
         $($coord:expr),*
         $(,)?
     ) => {
-        $crate::LineString(
+        $crate::LineString::new(
             <[_]>::into_vec(
                 ::std::boxed::Box::new(
                     [$($coord), *]
@@ -296,7 +310,7 @@ macro_rules! polygon {
 mod test {
     #[test]
     fn test_point() {
-        let p = point!(x: 1.2, y: 3.4);
+        let p = point! { x: 1.2, y: 3.4 };
         assert_eq!(p.x(), 1.2);
         assert_eq!(p.y(), 3.4);